code review stage (a)
