+++ src/services/prisma-datastore/prismaToSparrow.ts
@@ -2,7 +2,7 @@ import Prisma from "@prisma/client";
 import Gateway from "../alpha-models/Gateway";
 import Node from "../alpha-models/Node";
 import { NodeSensorTypeNames } from "../DomainModel";
-import { SignalStrengths } from "../alpha-models/SignalStrengths";
+import { LoraSignalMetricsToSignalStrengths } from "../alpha-models/SignalStrengths";
 
 export type SensorLatestReadings = (Prisma.Sensor & {
   latest: Prisma.Reading | null;
@@ -19,59 +19,12 @@ export type GatewayWithLatestReadings = Prisma.Gateway &
   LatestReadingSourceReadings;
 export type NodeWithLatestReadings = Prisma.Node & LatestReadingSourceReadings;
 
-interface LoRaSignalMetrics {
-  rssi: number;
-  snr?: number;
-}
-
-function loraSignalMetricsToSignalStrengths(
-  metrics?: LoRaSignalMetrics
-): SignalStrengths {
-  if (!metrics) {
-    return "N/A";
-  }
-
-  const bars = loraSignalMetricsToBars(metrics);
-  return barsToSignalStrengths(bars);
-}
-
-function loraSignalMetricsToBars(metrics: LoRaSignalMetrics): number {
-  const bars = rssiToBars(metrics.rssi);
-  const snrOffset = metrics.snr !== undefined ? snrBarsOffset(metrics.snr) : 0;
-  return bars + snrOffset;
-}
-
-function snrBarsOffset(snr: number): number {
-  if (snr < -10) {
-    return -0.5;
-  } else if (snr < 0) {
-    return -0.25;
-  }
-  return 0;
-}
-
-function rssiToBars(rssi: number): number {
-  return (rssi + 120) / 30;
-}
-
-function barsToSignalStrengths(bars: number): SignalStrengths {
-  if (bars < 0.1) {
-    return "0";
-  } else if (bars < 1.5) {
-    return "1";
-  } else if (bars < 2.5) {
-    return "2";
-  } else if (bars < 3.5) {
-    return "3";
-  } else return "4";
-}
-
 /**
  * Converts a prisma gateway to the old domain model.
  * @param gw
  * @returns
  */
-function sparrowGatewayFromPrismaGateway(
+export function sparrowGatewayFromPrismaGateway(
   pGateway: Prisma.Gateway,
   voltage: number
 ): Gateway {
@@ -127,7 +80,7 @@ function downscale(value: number | undefined, scale: number) {
   if (value !== undefined) return value / scale;
 }
 
-function sparrowNodeFromPrismaNode(
+export function sparrowNodeFromPrismaNode(
   gatewayUID: string,
   prismaNode: NodeWithLatestReadings
 ): Node {
@@ -139,7 +92,7 @@ function sparrowNodeFromPrismaNode(
     location: prismaNode.locationName || "",
     gatewayUID,
     lastActivity: prismaNode.lastSeenAt?.toString() || "",
-    bars: loraSignalMetricsToSignalStrengths(),
+    bars: LoraSignalMetricsToSignalStrengths(),
     temperature: asNumber(findReading(map, NodeSensorTypeNames.TEMPERATURE)),
     humidity: asNumber(findReading(map, NodeSensorTypeNames.HUMIDITY)),
     // todo - scaling should be driven by the sensor reading schema, but for now hard-coding to 100
@@ -156,7 +109,7 @@ function sparrowNodeFromPrismaNode(
     findReading(map, NodeSensorTypeNames.LORA_SIGNAL_STRENGTH)
   );
   if (rssi !== undefined) {
-    node.bars = loraSignalMetricsToSignalStrengths({ rssi });
+    node.bars = LoraSignalMetricsToSignalStrengths({ rssi });
   }
 
   if (node.temperature === undefined) delete node.temperature;
@@ -168,9 +121,6 @@ function sparrowNodeFromPrismaNode(
 
   return node as Node;
 }
-
-// todo - what's the purpose behind this construct?
-export { sparrowGatewayFromPrismaGateway, sparrowNodeFromPrismaNode };
 
 export const DEFAULT = {
   sparrowGatewayFromPrismaGateway,
